CCFI/Sonder/requirements.txt
invoke==1.7.1
hug==2.6.0
pandas==1.3.3
mysql-connector-python==8.0.26